+++ Cpp/custom/collected/src/Flow_network.h
@@ -13,9 +13,9 @@ public:
 
     Flow_network(In& in);
 
-    inline int num_vertices() const noexcept { return _num_vertices; }
+    inline unsigned long num_vertices() const noexcept { return _num_vertices; }
 
-    inline int num_edges() const noexcept { return _num_edges; }
+    inline unsigned long num_edges() const noexcept { return _num_edges; }
 
     void add_edge(Flow_edge& e);
 
@@ -28,8 +28,8 @@ public:
     std::string to_string();
 
 private:
-    const int _num_vertices;
-    int _num_edges;
+    const unsigned long _num_vertices;
+    unsigned long _num_edges;
     std::vector<std::vector<Flow_edge>> _adjacency_lists;
 
     void _validate_vertex(int v);
